// src/types.ts
export interface AppConfig {
      registrationWebhook: string;
      agentConfigWebhook: string;
      agentUpdateWebhook: string;
      qrWebhook: string;
      openaiApiKey: string;
      geminiApiKey: string;
      baserowApiUrl: string;
      baserowToken: string;
      baserowTableId: string;
      baserowNameColumn: string;
      baserowEmailColumn: string;
      baserowWhatsappColumn: string;
      baserowBusinessColumn: string;
      baserowDescriptionColumn: string;
      baserowFunctionColumn: string;
      baserowPasswordColumn: string;
    }